Removed BitStream. Moved PartitionedStream to OwlCore.IO. Updated changelog.
