+++ src/IO/StreamPartitionStates.cs
@@ -1,0 +1,14 @@
+﻿using System;
+
+namespace OwlCore.IO;
+
+/// <summary>
+/// Indicates the various possible states for a <see cref="StreamPartition"/>.
+/// </summary>
+[Flags]
+public enum StreamPartitionState : byte
+{
+    None = 0,
+    Deleted = 1,
+    Fragmented = 2,
+}
